api/requirements.txt
flask==2.2.2
protobuf==3.20.3
google-cloud-bigquery==3.5.0
uuid==1.30